% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{Generate_Case_Full_Data}
\alias{Generate_Case_Full_Data}
\title{Simulation Case Dataset Generator}
\usage{
Generate_Case_Full_Data(case, n, p, sigmax = NULL)
}
\arguments{
\item{case}{integer. Case number for simulation.}

\item{n}{integer. Sample size of full data.}

\item{p}{integer. Number of predictors excluding intercept.}

\item{sigmax}{matrix. Covariance matrix of predicotrs, default is auto-correlated.}
}
\description{
Simulation Case Dataset Generator
}
