% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{get_Iboss}
\alias{get_Iboss}
\title{Get IBOSS Subsample from user input data}
\usage{
get_Iboss(data, k, y_pos = 0, response.included = F)
}
\arguments{
\item{response.included}{logical. If TRUE, repsonse vector is included in dataset. default is false.}
}
\description{
Get IBOSS Subsample from user input data
}
