% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{get_mse}
\alias{get_mse}
\title{calculating Mean Squared Errors (MSE)}
\usage{
get_mse(case, n, k, p, re = 1000, compare = F)
}
\arguments{
\item{re}{integer, time of repetitions}

\item{caseData}{a list caseData has to be a list consists of 
design matrix: X(intercept included)
response vector: Y 
true parameter: beta
total number of obs(possible) : N
case number: case
number of predictors: dim(beta) - 1}
}
\description{
calculating Mean Squared Errors (MSE)
}
