% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{getIBOSS0}
\alias{getIBOSS0}
\title{Function to get IBOSS sample only}
\usage{
getIBOSS0(data, k, y_pos = 0)
}
\arguments{
\item{data}{dataframe or data.table}

\item{k}{integer, desired subsample size}

\item{y_pos}{integer, position of response column}
}
\description{
Function to get IBOSS sample only
}
