% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getIdx}
\alias{getIdx}
\title{Get IBOSS sample from one dimension}
\usage{
getIdx(x, f, rr, max)
}
\arguments{
\item{x}{current remainder of one column data frame or vector}

\item{f}{full one column dataframe or vector}

\item{rr}{integer, number of selected samples, rr = k / d / 2}

\item{max}{integer, total subsample size, k}
}
\value{
vector of selected indices in full dataframe.
}
\description{
Get IBOSS sample from one dimension
}
