% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iboss.r
\name{iboss.od}
\alias{iboss.od}
\title{The IBOSS method}
\usage{
iboss.od(Z, Y, k, int.adj = "TRUE")
}
\arguments{
\item{Z}{the input covariate matrix or covariate vector}

\item{Y}{the response vector}

\item{k}{the subdata size}

\item{int.adj}{whether to calculate the adjusted estimate of the intercept. It is \code{TRUE} by default.}
}
\description{
This function implements the IBOSS method for the input covariate Z and response vector Y. It returns a list with elements: beta, the least squares estimate based on the subdata; se, the standard errors; sigma, variance estimate for the error term, index, index of the subdata.
}
\examples{
library(IBOSS)
library(mvtnorm)
beta.true  <- rep(1, 51)
d <- length(beta.true) - 1
corr  <- 0.5
sigmax <- matrix(corr, d, d) + diag(1-corr, d)
n <- 5000
k <- 100
set.seed(0)
X  <- rmvt(n, sigmax, 2)
mu  <- beta.true[1] + c(X \%*\% beta.true[-1])
Y  <- mu + rnorm(n, 0, 3)
fit <- iboss.od(X, Y, k)
beta.od <- fit$beta
beta.odia <- fit$beta0.adj
}

