% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{runIbossExample}
\alias{runIbossExample}
\title{Wapper of pre-designed examples.}
\usage{
runIbossExample(ex = NULL)
}
\arguments{
\item{ex}{integer, menu based.}
}
\value{
MSEs and plots
}
\description{
Wapper of pre-designed examples.
}
\examples{
\dontrun{
# use menu based selector
runIbossExample()
# or input case number by user
runIbossExample(1)
}
}
