% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demoIboss}
\alias{demoIboss}
\title{Customized Simulation Functions, returns MSE and plots}
\usage{
demoIboss(case = NULL, n = NULL, k = NULL, p = NULL, rept = NULL,
  compare = TRUE, mse.plot = TRUE)
}
\description{
Customized Simulation Functions, returns MSE and plots
}
