% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{getMSE}
\alias{getMSE}
\title{calculating Mean Squared Errors (MSE)}
\usage{
getMSE(case, n, k, p, rept = NULL, compare = F)
}
\arguments{
\item{case}{integer. Distribution to generate full data.}

\item{n}{integer. Full data size.}

\item{k}{integer. Subsample size.}

\item{p}{integer. Number of predictors.}

\item{rept}{integer. Repeat times, defualt is 100.}

\item{compare}{logical. If True, output comparisons to other subsampling methods; Otherwise, output iboss sampling results only.}
}
\description{
calculating Mean Squared Errors (MSE)
}
